refactor: New schema, AI module consolidation, and workflow restructure
Schema changes (fresh DB):
- Rename steps_completed→current_phase, steps_total→total_phases
- Replace column enum with done_at timestamp
- Add setup_steps JSON map for tracking setup progress
- New ai_sessions table (claude_session_id, worktree_path)
- Messages belong to ai_sessions instead of workflow_sessions
- Add Tailwind source for workflow module CSS class detection

AI module consolidation:
- Merge Messages, AiSessions, and generate_title into Destila.AI context
- Rename AI.Session → AI.ClaudeSession (GenServer wrapping ClaudeCode)
- New AI.Session schema (table: ai_sessions)
- New AI.Message schema (moved from Messages.Message)
- AI.process_message/2 replaces Messages.process/2

Workflow restructure:
- PromptChoreTaskWorkflow: phases/0 returns [{Module, opts}] tuples (6 phases)
- Workflow metadata: label/0, description/0, icon/0, icon_class/0
- SetupCoordinator replaces Destila.Setup (same atomic CAS pattern)
- Workers updated: setup_steps JSON, AI session in AiConversationPhase
- Delete static workflows (PromptNewProject, ImplementGenericPrompt)
