+++ assets/css/app.css
@@ -5,6 +5,7 @@
 @source "../css";
 @source "../js";
 @source "../../lib/destila_web";
+@source "../../lib/destila/workflows";
 
 /* A Tailwind plugin that makes "hero-#{ICON}" classes available.
    The heroicons installation itself is managed by your mix.exs */
