+++ lib/destila/workers/setup_worker.ex
@@ -1,7 +1,7 @@
 defmodule Destila.Workers.SetupWorker do
   use Oban.Worker, queue: :setup, max_attempts: 3
 
-  alias Destila.{Git, Messages, WorkflowSessions}
+  alias Destila.{Git, WorkflowSessions}
 
   @impl Oban.Worker
   def perform(%Oban.Job{args: %{"workflow_session_id" => workflow_session_id}}) do
@@ -12,8 +12,8 @@ defmodule Destila.Workers.SetupWorker do
         do: Destila.Projects.get_project(workflow_session.project_id)
 
     with :ok <- sync_repo(workflow_session, project),
-         :ok <- create_worktree(workflow_session, project),
-         :ok <- start_ai_session_and_trigger(workflow_session) do
+         :ok <- create_worktree(workflow_session, project) do
+      Destila.Workflows.SetupCoordinator.maybe_advance_setup(workflow_session_id)
       :ok
     end
   end
@@ -23,33 +23,28 @@ defmodule Destila.Workers.SetupWorker do
   defp sync_repo(workflow_session, project) do
     cond do
       project.local_folder && project.local_folder != "" ->
-        broadcast_step(
-          workflow_session.id,
-          "repo_sync",
-          "in_progress",
-          "Pulling latest changes..."
-        )
+        update_setup_step(workflow_session.id, "repo_sync", "in_progress")
 
         case Git.pull(project.local_folder) do
           {:ok, _} ->
-            broadcast_step(workflow_session.id, "repo_sync", "completed", "Repository up to date")
+            update_setup_step(workflow_session.id, "repo_sync", "completed")
             :ok
 
           {:error, reason} ->
-            broadcast_step(workflow_session.id, "repo_sync", "failed", reason)
+            update_setup_step(workflow_session.id, "repo_sync", "failed", reason)
             {:error, reason}
         end
 
       project.git_repo_url && project.git_repo_url != "" ->
-        broadcast_step(workflow_session.id, "repo_sync", "in_progress", "Syncing repository...")
+        update_setup_step(workflow_session.id, "repo_sync", "in_progress")
 
         with {:ok, path} <- Git.effective_local_folder(project),
              {:ok, _} <- Git.pull(path) do
-          broadcast_step(workflow_session.id, "repo_sync", "completed", "Repository up to date")
+          update_setup_step(workflow_session.id, "repo_sync", "completed")
           :ok
         else
           {:error, reason} ->
-            broadcast_step(workflow_session.id, "repo_sync", "failed", reason)
+            update_setup_step(workflow_session.id, "repo_sync", "failed", reason)
             {:error, reason}
         end
 
@@ -66,75 +61,45 @@ defmodule Destila.Workers.SetupWorker do
         worktree_path = Path.join([local_folder, ".claude", "worktrees", workflow_session.id])
 
         if Git.worktree_exists?(worktree_path) do
-          WorkflowSessions.update_workflow_session(workflow_session.id, %{
-            worktree_path: worktree_path
+          update_setup_step(workflow_session.id, "worktree", "completed", nil, %{
+            "worktree_path" => worktree_path
           })
 
-          broadcast_step(workflow_session.id, "worktree", "completed", "Worktree ready")
           :ok
         else
-          broadcast_step(workflow_session.id, "worktree", "in_progress", "Creating worktree...")
+          update_setup_step(workflow_session.id, "worktree", "in_progress")
 
           case Git.worktree_add(local_folder, worktree_path, workflow_session.id) do
             {:ok, _} ->
-              WorkflowSessions.update_workflow_session(workflow_session.id, %{
-                worktree_path: worktree_path
+              update_setup_step(workflow_session.id, "worktree", "completed", nil, %{
+                "worktree_path" => worktree_path
               })
 
-              broadcast_step(workflow_session.id, "worktree", "completed", "Worktree ready")
               :ok
 
             {:error, reason} ->
-              broadcast_step(workflow_session.id, "worktree", "failed", reason)
+              update_setup_step(workflow_session.id, "worktree", "failed", reason)
               {:error, reason}
           end
         end
 
       {:error, reason} ->
-        broadcast_step(workflow_session.id, "worktree", "failed", reason)
+        update_setup_step(workflow_session.id, "worktree", "failed", reason)
         {:error, reason}
     end
   end
 
-  defp start_ai_session_and_trigger(workflow_session) do
-    broadcast_step(workflow_session.id, "ai_session", "in_progress", "Starting AI session...")
+  defp update_setup_step(workflow_session_id, step, status, error \\ nil, extra \\ %{}) do
+    ws = WorkflowSessions.get_workflow_session!(workflow_session_id)
 
-    workflow_session = WorkflowSessions.get_workflow_session!(workflow_session.id)
+    step_value =
+      %{"status" => status}
+      |> then(fn v -> if error, do: Map.put(v, "error", sanitize_error(error)), else: v end)
+      |> Map.merge(extra)
 
-    session_opts =
-      Destila.AI.Session.session_opts_for_workflow(
-        workflow_session,
-        workflow_session.steps_completed,
-        timeout_ms: :timer.minutes(15)
-      )
+    setup_steps = Map.put(ws.setup_steps || %{}, step, step_value)
 
-    case Destila.AI.Session.for_workflow_session(workflow_session.id, session_opts) do
-      {:ok, _session} ->
-        broadcast_step(workflow_session.id, "ai_session", "completed", "AI session ready")
-
-        Destila.Setup.maybe_finish_phase0(workflow_session.id)
-        :ok
-
-      {:error, reason} ->
-        broadcast_step(workflow_session.id, "ai_session", "failed", inspect(reason))
-        {:error, reason}
-    end
-  end
-
-  defp broadcast_step(workflow_session_id, step, status, content) do
-    content =
-      if status == "failed" do
-        sanitize_error(content)
-      else
-        content
-      end
-
-    Messages.create_message(workflow_session_id, %{
-      role: :system,
-      content: content,
-      raw_response: %{"setup_step" => step, "status" => status},
-      phase: 0
-    })
+    WorkflowSessions.update_workflow_session(ws, %{setup_steps: setup_steps})
   end
 
   defp sanitize_error(message) when is_binary(message) do
