+++ lib/destila/workflow_sessions/workflow_session.ex
@@ -6,28 +6,23 @@ defmodule Destila.WorkflowSessions.WorkflowSession do
   @foreign_key_type :binary_id
   schema "workflow_sessions" do
     field(:title, :string, default: "Untitled Session")
+    field(:workflow_type, Ecto.Enum, values: [:prompt_chore_task])
 
-    field(:workflow_type, Ecto.Enum,
-      values: [:prompt_chore_task, :prompt_new_project, :implement_generic_prompt]
-    )
-
-    field(:column, Ecto.Enum, values: [:request, :distill, :done])
-
-    field(:steps_completed, :integer, default: 0)
-    field(:steps_total, :integer, default: 4)
+    field(:current_phase, :integer, default: 1)
+    field(:total_phases, :integer)
 
     field(:phase_status, Ecto.Enum,
       values: [:setup, :generating, :conversing, :advance_suggested]
     )
 
     field(:title_generating, :boolean, default: false)
-    field(:ai_session_id, :string)
-    field(:worktree_path, :string)
+    field(:setup_steps, :map, default: %{})
     field(:position, :integer)
+    field(:done_at, :utc_datetime)
     field(:archived_at, :utc_datetime)
 
     belongs_to(:project, Destila.Projects.Project)
-    has_many(:messages, Destila.Messages.Message)
+    has_many(:ai_sessions, Destila.AI.Session)
 
     timestamps(type: :utc_datetime)
   end
@@ -38,16 +33,17 @@ defmodule Destila.WorkflowSessions.WorkflowSession do
       :title,
       :workflow_type,
       :project_id,
-      :column,
-      :steps_completed,
-      :steps_total,
+      :done_at,
+      :current_phase,
+      :total_phases,
       :phase_status,
       :title_generating,
-      :ai_session_id,
-      :worktree_path,
+      :setup_steps,
       :position,
       :archived_at
     ])
-    |> validate_required([:title, :workflow_type, :column])
+    |> validate_required([:title, :workflow_type])
   end
+
+  def done?(%__MODULE__{done_at: done_at}), do: not is_nil(done_at)
 end
