+++ lib/destila/workflows.ex
@@ -5,31 +5,34 @@ defmodule Destila.Workflows do
   """
 
   @workflow_modules %{
-    prompt_chore_task: Destila.Workflows.PromptChoreTaskWorkflow,
-    prompt_new_project: Destila.Workflows.PromptNewProjectWorkflow,
-    implement_generic_prompt: Destila.Workflows.ImplementGenericPromptWorkflow
+    prompt_chore_task: Destila.Workflows.PromptChoreTaskWorkflow
   }
 
-  @doc """
-  Returns the workflow module for a given workflow type.
-  """
   def workflow_module(workflow_type) do
     Map.fetch!(@workflow_modules, workflow_type)
   end
 
-  def steps(workflow_type), do: workflow_module(workflow_type).steps()
-  def total_steps(workflow_type), do: workflow_module(workflow_type).total_steps()
+  def workflow_types, do: Map.keys(@workflow_modules)
+
+  def workflow_type_metadata do
+    Enum.map(@workflow_modules, fn {type, mod} ->
+      %{
+        type: type,
+        label: mod.label(),
+        description: mod.description(),
+        icon: mod.icon(),
+        icon_class: mod.icon_class()
+      }
+    end)
+  end
+
+  def phases(workflow_type), do: workflow_module(workflow_type).phases()
+  def total_phases(workflow_type), do: workflow_module(workflow_type).total_phases()
   def phase_name(workflow_type, phase), do: workflow_module(workflow_type).phase_name(phase)
   def phase_columns(workflow_type), do: workflow_module(workflow_type).phase_columns()
   def default_title(workflow_type), do: workflow_module(workflow_type).default_title()
   def completion_message(workflow_type), do: workflow_module(workflow_type).completion_message()
 
-  @doc """
-  Returns the session strategy for a given workflow type and phase
-  as a normalized `{action, opts}` tuple.
-
-  Defaults to `{:resume, []}` for workflow types that don't define `session_strategy/1`.
-  """
   def session_strategy(workflow_type, phase) do
     module = workflow_module(workflow_type)
